pyproject.toml
[project]
name = "hdr_fuse"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "SiTinc", email = "23432137+Tinnci@users.noreply.github.com"},
]
dependencies = [
    "opencv-python>=4.10.0.84",
    "numpy>=2.1.2",
    "pillow>=11.0.0",
    "PyQt5>=5.15.11",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
